requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Model backbone
timm>=0.9.16

# Augmentation
albumentations>=1.3.1
opencv-python-headless>=4.9.0

# Metrics & logging
torchmetrics>=1.3.0
wandb>=0.16.0

# Config & utilities
pyyaml>=6.0.1
numpy>=1.26.0
Pillow>=10.2.0
matplotlib>=3.8.0
tqdm>=4.66.0
scikit-learn>=1.4.0

# Optional: AUC-maximisation loss
# libauc>=1.3.0
